Move logic from LayoutMatcher.cs to ZoneWatcher.cs
